LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/cssnumericvalue-multiply-two-types.tentative-expected.txt


FAIL Multiplying two types with different non-null percent hints throws TypeError assert_throws_js: function "() => new CSSMathProduct(a, b)" did not throw
FAIL Multiplying two types with same base types adds exponents assert_equals: length expected 2 but got 0
FAIL Multiplying two types with different base types adds exponents assert_equals: length expected 1 but got 0
FAIL Multiplying two types respects the sign of the exponents assert_equals: length expected 1 but got 0
FAIL Multiplying a type with no exponents is a no-op assert_equals: length expected 1 but got 0
FAIL Multiplying a type with percent hint applies the percent hint assert_equals: length expected 2 but got 0
FAIL Multiplying two types with same percent hint applies the percent hint assert_equals: length expected 2 but got 0
